+++ test/TestMatrix.cpp
@@ -6,12 +6,12 @@
 
 void testMatrix() {
     Matrix<uint8_t, 2, 3> mat{};
-    for (int i = 0; i < 2; i++) for (int j = 0; j < 3; j++) assert(mat[i][j] == 0);
+    for (size_t i = 0; i < 2; i++) for (size_t j = 0; j < 3; j++) assert(mat[i][j] == 0);
 
     Matrix<uint8_t, 3, 4> mat2 = Matrix<uint8_t, 3, 4>::identity();
     Matrix<uint8_t, 2, 4> product = mat * mat2;
     Matrix<uint8_t, 2, 4> z = Matrix<uint8_t, 2, 4>::zeros();
-    for (int i = 0; i < 2; i++) for (int j = 0; j < 4; j++) assert(product[i][j] == z[i][j]);
+    for (size_t i = 0; i < 2; i++) for (size_t j = 0; j < 4; j++) assert(product[i][j] == z[i][j]);
 
     Matrix<int, 1, 2> d_mat = Matrix<int, 1, 2>::identity();
     Matrix<int, 1, 2> s_mat = d_mat.applyFunc<int>([] (const int &num) { return num * num; });
